WIP: First stab at compiling patterns
Each case in my AST has a syntax class, and each syntax class has an
info attribute. That info attribute will eventually contain a function
to compile an instance of that case in the AST as a template, but for
now all it contains is a function to compile an instance of that case
as a pattern.

Making an info attribute for a combination means that I'll need the
info attributes for its parts. To get the info attribute for an
arbitrary part, I'll use the function bound to getinfo-id.