WIP: Add classes and shapes for primitive ops; fix nonterminal bug
Primitive ops will be parameterized over the c/e/k classes and shapes
and not over some toplevel shape/class (given the current
representation, I don't know how we would even generate it).

Classes for nonterminals need to know what their alternatives' shapes
are---otherwise, the rule for checking that the value in the function
continuation is a lambda will only check that it's a value. Because of
the recursive nature of classes, I think we'll need to construct them
in a similar way to how we construct shapes (possibly by constructing
the shape and class maps at the same time).