Translate IR to strings of Python code
Also, I made ir:class-def and classes a little more in sync: when a
class has #f for its fields, the class-def too should have #f for its
fields. This way, an ir:make that sends #f for its args is consistent
with the class-def having #f. None of this matters that much, since we
don't do anything special in those cases (e.g. by reusing a singleton
instance); but at least we're consistent.