Print classes in .py in an order that Python won't obviously scream
There's still a possibility for things to not be defined in the right
order if e.g. the production for variables is written above the
production for expressions---there, the current code would spit out
the class definition for var, which is a subclass of e, before the
class definition for e. Maybe it is worth it to write a proper sort..