Add a side-condition like thing; add lexical scope to our LC example
The #:where clauses (I've been calling them side-conditions) are very,
very simple. They are a glorified sequence of pattern matches and, due
to how we compile the rest of the system, they are not suitable to
distinguish overlapping cases. For example, you cannot use #:where to
distinguish rules like below

[(v env k) -> (...)
 #:where (lam x e) v]
[(v env k) -> (...)
 #:where (lam x e) n]

Because of that, I don't think the keyword #:where gives the right
intuition..