Use language struct to parse arbitrary terms/patterns
Also: this commit moves template parsing towards actual typechecking,
where we produce a type for each term.

Each form implements a separate piece of the overall language's
pattern/term parser (type checker). Each form's piece corresponds to a
case in the typing judgment I sketched out, but the judgment says that
we need a way for the form to recursively call the judgment on its
subpatterns/subterms. This sort of recursion has been tricky to thread
through in past designs, but the least-bad way I came up with was to
make forms recur through the language struct.