New frontend design with explicit parsing stage
The new design comes with a proper definition of the AST we'll walk
over. Since I'm still not comfortable generating a grammar-specific
AST (e.g. one where a function application node in the AST is
represented by a struct with two fields), I'm going with a general
one.

I should clarify this in the code, but I'm not sure how: symbol AST
nodes represent nonterminals that are not part of a compound piece of
syntax like (cons e e). In the syntax for cons expressions, for
example, the part in function position (the literal cons) would not be
a symbol AST node.

I'm also uncertain how to clean up the -compound-of syntax class. One
thing I'd like to figure out is how to not produce a dummy value in
the base case, but figuring that out will have to wait.

Note: code related to typechecking is only a sketch; there's nothing
using it at the moment and it will be replaced in later commits.

I'm calling the programmatic hooks into the language /primitives/. The
interface is still rough so that too is bound to change.