Kill rightward drift in step->methods; refactor typechecking a bit
So that typechecking a step can use a slightly more uniform
interface (via tc-ast*s) I introduced the ast* datatype. I don't have
a good handle on exactly what it represents (in the world of ideas)
but it encodes expected type information so that a sequence of ast
nodes (after being turned into ast* nodes) can be typechecked easily.

Also, I made it so the nasty build/ rules that cat files together
properly fail when one of the commands fails.