Add native integers as a primitive; convert fact7 to barely use it
When I tried to compute factorial of 10 using this implementation,
running it on PyPy finished in about 20 seconds. When I ran the
interpreter+program through RPython I had to stop the resulting binary
at 4m30s; by that point, activity monitor said it was using 17GB of
memory. Running the corresponding Racket code finished in a couple
seconds.