Replace eq dunder with custom eq
RPython doesn't use/call a lot of the dunders on user-defined classes,
so we have to implement our own.

This was why my RPython-translated factorial binaries were allocating
so much memory: the eq test was always false, which meant iszero
always failed. My base cases all relied on iszero working, so when it
didn't, the functions that used it would just infinite loop and
allocate as quickly as possible.