Hack together specifying initial config; hack together initial basis
This required changing what the primitive/metafunctions for succ,
etc. did. Because the functions in the initial basis use the
metafunctions with a variable as an argument, the primitives can
either a) do the variable lookup/closure destructuring themselves or
b) act like any other expression and participate in the CEK calling
convention. I went with the latter even though it's more heavyweight
than we need (i.e. option a would avoid a round trip or two through
the main loop).