Use old-school CEK semantics to only pass values on the stack
Because the ignore terminal is used in the control string, we have to
make sure that terminals are compiled to a singleton. It's
specifically used when primitive expressions return _ignore_sing as
their control string.

Because closures are now values and basic constants are no longer
wrapped in a closure, implementing primitives is easier. Specifically,
if they have a variable and need its value, they can just consult an
environment and not care whether the value is a basic constant that
needs to be pulled out of a closure or whether it's a lambda that
needs its closure. Because we can now write primitives that can
easily access a variable's value, simple primitives no longer need
separate continuations to evaluate their arguments.