Replace the ignore exp/ret continuation with a single value wrapper
The value wrapper lets us pass values through the control string
without throwing the JIT off. As long as there's a (compile-time)
fixed amount of these wrappers then they can appear in the greens
(control string). The trade off is that (for now?), to return a value,
we need to call a metafunction that reuses the wrapper instance(s).

It also means we can avoid the additional complexity to make terminals
singletons.