Fix the rules for let; fix failing tests
I don't know why the previous commit required me to comment out let,
but this fix was sufficient. Considering all of the runtime code that
I touched did not change in the last commit, I'm really surprised that
it became an issue. Maybe it has to do with the two different kinds of
continuations that the last commit introduced?