Replace let with let-values; add multiple values return
The values form in this implementation is weird---since values is
supposed to be a function in the initial basis, I chose to implement
that as a variadic function whose body has a form (maybe poorly
named?) values that expects a single argument which should be bound to
a vlist in the environment of the function, i.e. the values form needs
to get the name of the rest args.