Final functions needed to get the fib benchmark to run
To run the fib benchmark when it's already been created by
pycket-bench's run-racket binary, use these commands:
    BENCH=~/projects/pycket-bench/CrossBenchmarks/racket/fib-nothing.rkt
    raco expand $BENCH | make quicktest

I added a hack that pares down some of the expansion. This is to
maintain whatever invariant is tied to the GCBitmaskTooLong
error... Once we have have a proper parser in the code we generate,
then we can avoid this issue more-correctly. For now, though, this
hack will do.

Also, fprintf doesn't work properly. It too is a hack.