Pull (Python) interpreter code into separate directory; fix makefile
This took some mangling of names to make things work with modules. I
initially tried to just insert liberal uses of 'from mumble import *'
but that had issues with variables not being defined at the right
time (for example, the calls to mkvariable inside of program.py).

I'm not entirely happy with the changes, but this isn't the end goal;
this is part of the larger goal of having a simplified parser in the
binary so that we can run more than one program per binary.