Remove the ret continuation and replace it with the ret exp
The thought was that the ret continuation's interpret rule was
spending too much time trying to find which rule applied. So, to get
around that, I changed the semantics a little and modified the
produced machine.py to make the interpret rule for the ret exp
immediately dispatch to its continuation.

This didn't seem to make the fib example any faster; the resulting
binary still took about 100x longer than the Pycket/Racket
versions. So, I'm now assuming that the slowdown wasn't due to the
slow backtracking in the ret continuation.

These changes are not necessarily correct, but I think they were
sufficient to check the hypothesis. Concretely, running the produced
binary will not exit normally but it will exit with an error for wrong
number of arguments to a function.