When looking up a variable, promote the next env (if nonvirtual)
ATM I don't really know why this helps so much/removes so much of the
runtime---the environments that we repeatedly walk through to look up
bindings in the "basis" (e.g. -, <, and + in the fib benchmark) aren't
virtual but I thought those were where we spent most of our time. I'll
have to compare the traces later to figure this out.