Bind module definitions at once, in order; ExtendedEnv immutable
I changed the module bindings to happen at once so that I could use a
different representation of extended environments (one that put
multiple binding pairs in a single environment instance) but I didn't
know how to optimize it.

Thankfully, I had a little bit of insight on how to optimize the
simpler ExtendedEnv traces---the fields of the ExtendedEnv need to be
immutable to remove the costly parts of the trace that walk up the
chain of environments. With those changes, I'm now within 10x of
Racket and the cost of larger environments with irrelevant
bindings (like the suite-wide metadata in CrossBenchmarks) is much
smaller.

I didn't roll back part of the semantics' changes, the parts that bind
the definitions at once (while still keeping the in order part),
because this seems a little cleaner.