Don't use jit.isvirtual to determine when we promote environments
The docs say that jit.isvirtual is conservative, but what they don't
say is what conservative means. Does it mean that isvirtual will say
some non-virtual thing is virtual? or does it mean that isvirtual will
say some virtual thing is not virtual? Unfortunately, experience/some
source files that use phrasing "is likely virtual" makes me believe it
uses the latter meaning of conservative, making it not a safe way to
guard promoting.

In our case, we don't need a general isvirtual and can instead
approximate it: if an environment was allocated while tracing, we
don't promote it. All virtual objects are allocated while tracing, so
I think this approximation is complete; but, not all objects allocated
while tracing are virtual (e.g. those that escape), which is why I
think this isn't a sound approximation.

The only way I could identify whether we were tracing or not was using
jit.current_trace_length(). It at least had a docstring so I'm hoping
it stays supported.

Finally, I don't know if the EmptyEnv/Env classes need the
_immutable_fields_ attributes, but I haven't yet looked into it.