Make variable equality checks constant time
This gave a small improvement to some of the benchmarks (fib, sum were
maybe < 5% improvement) but a moderate improvement to others (mbrot,
nqueens were closer to > 10% improvement).

I have no idea why the type checker balked at the guard inside
Float's __init__, but I had to remove it. Nothing seems to have broken
outside of that, but I'm puzzled...