Promote formals; remove unused/overspecialized promote attr
Lists of variables should only be constructed at (roughly) parse time,
so we can promote those--they should never be constructed on
trace. This means that we can elide constructing the variable to
offset dictionary.

Similarly, we can annotate the values list immutable and not just its
field (but we have to take a copy of it, apparently).