Linear search in the environment instead of using a dictionary
Pycket does this, which is why I initially tried it (knowing that
nothing else appeared to be helping improve variable lookup
time). They do a linear search over a list of atoms which is basically
how I have to search over a varl.

Instead of storing the mapping in a dictionary that is somewhat opaque
to the JIT, iterating over an immutable class and unrolling the loop
that does the iteration optimizes better.