Replace some calls to extend1 with extend; add metafun extendcells
I'm trying to remove calls to extend1 so that I more uniformly use the
MultiExtendedEnv throughout. I introduced a new metafun (extendcells)
for one call site, and for the other two cases I'm just using the
normal extend function on singleton lists.

I still need to rework extendrest to not call extend1, but the changes
in lc.rkt is all I feel like doing for now.