Try again at speculating
Now (like Pycket) the environment structure is stored at the
application sites, and it looks like the JIT can constant fold
operations that rely on it.

Note: without the assert in Env1::__init__, the type for an Env's e
attribute is inferred as cl_env, when we need it to be inferred as Env

(I also got around to making the classes extendable. Boy was that easy!)