Dynamic loop detection using Pycket's CallGraph class
But with hacks! And with annotations that for some unknown reason are
required! This is split between the compiler and the runtime because
I couldn't fully do it in the runtime (I couldn't get an instance
field added to a class via __extend__, only by a metaclass, and then
that didn't cooperate with __extend__ so I bent and made these changes
in the compiler).

The file interpreter/pycket/callgraph.py is Pycket's current
callgraph.py, which was changed in commit ad21e0ee