Distinguish between promotable and non-promotable cells
Promoting the closure for top-level definitions seems to help all of
the benchmarks, and that's easy to do by splitting the Cell class two
subclasses: one whose content we promote and one whose content we
don't.

As long as programmers don't reinstate the continuation for define's
this should be a good improvement.