Add/update primitives to support more benchmarks
With small tweaks (e.g. around quoting) I can now run the benchmarks
added to run-cross.rkt.

Why is the definition of + so messy? The old version was limited to
only binary uses of +, but to support some benchmarks that require
more than 2 args (e.g. 3 args) I added a general case. But, the
benchmarks that needed only 2 args slowed down quite a bit; fib, for
example, went from about 10 seconds to 20 seconds. So this good
enough for now but I don't like how ad-hoc it is.