A reasonably complete Impcore abstract machine+Racket impl
For now, much copypasta from the lc example. But I'm relatively happy
with doing the expansion part using a Racket language: it seems easy
enough to do something like Pycket's expand module when I have a #lang
for Impcore. And, for Impcore at least, once I have it as a #lang I
might as well make it one that executes correctly.

There's still some mess (alluded to in comments) with expand, but this
is just about done.

I added some hacks so that runtime.py only has to include the
metafunctions that a language actually uses. But since the LC language
always needs the box-related metafunctions at parse->JSON AST time,
all languages have to have those in runtime.py. Since the impcore
language needs println at parse->JSON AST time, that's also required
in runtime.py. And similarly all languages have to have the prim
parser-related functions (e.g. mkfloat) in their runtime.py

And in this mess, the change to parse.rkt fixes a bug in my
understanding of syntax-e/unsyntax-splicing. I expected the last
expression to be a (syntax) list of identifiers, which it is. And I
expected the expression before that to be a list of identifiers, but
it's a cons of an identifier and a syntax-pair

(define stx (list #'snacks #'sweets #'drinks))
(syntax-e #`(groceries #,@stx))

I also dropped Pycket's absurdly high trace_limit. I was getting
transition-errors that would occur immediately after a trace abort,
and I think that the limit being set so high was the cause (maybe such
a high value violates an invariant in how RPython recovers from
aborted traces). But I don't know for certain.

I started with a different begin continuation (still called seq) that
an arbitrary es but I believe that would accumulate seq continuations
that just held an esnil. It seems like the general strategy to not
accumulate unnecessary continuations is to ask at every transition
that pushes a continuation: do I *really* need to push a continuation
here? can I change my representation somewhere to avoid pushing it?