Let examples overlay their own Python code
I'm not yet sure how I want to disconnect the number code from the
booleans it may return (e.g. for the result of <), so for now I'm just
going to duplicate things. Then maybe some patterns will emerge.

But each overlay needs to contain a runtime.py that has a few things:
a a JitDriver bound as driver, a function from programs to answers
bound as run, and exceptions CEKMatchfailure, CEKUnlessFailure, and
CEKDone.