Get rid of spurious guard by moving if continuation code lower
When the if continuation code was at the top, there were mysterious
guards about the 0-ness of a return value. I noticed this first in the
traces of the sum benchmark, where there was a guard that the
accumulator was not zero. That guard was coming from the code for
dispatching on a ret continuation: its first case checked whether the
returned value was 0 or not, and then it would check whether the value
was being returned to a sel continuation. When that second check
failed, it would then go onto checking the next case. There were no
guards in the trace on the failed continuation check (the optimizer
sure does clean things up...) but the guard that the value was not
zero remained.

Even with a better pattern matcher, it seems like having an ordered
interpretation of these rules will always have an opportunity for
these non-obvious guards in a trace.