Impcore: Promote the function that we get out of an environment
In the LC version of sum we didn't have to do this so explicitly
because the closure for the main loop gets promoted. But in Impcore we
don't have the PromotableClosure code; instead, since the set of
functions in an Impcore program is small we can promote everything
that comes out of the function environment (If we had closures,
though, in the general case w could only promote the lambda part).

I had to change the function continuations a little bit---since for
the moment I can only call trypromote on v's I couldn't take apart the
fun.